+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 * Fixed the **generate-docs** to handle integration authentication parameter.
+* Added a validation to ensure that description and README do not contain the word 'Demisto'.
 
 # 1.3.9
 * Added a validation verifying that the pack's README.md file is not equal to pack description.